fix(ai): address code review feedback for Phase 4
- Issue 1 (Important): Add missing `refillRate` field in tokenBucket example
  for ArcjetDeniedError JSDoc. Required field was omitted, causing the
  example to be incomplete.

- Issue 2 (Minor): Fix protectAction @example to use real rule constructor
  (fixedWindow) instead of undefined `rateLimit`. Added all required config
  keys (maxRequests, windowSeconds) and imports so the example is
  conceptually compilable.

- Issue 3 (Minor): Extract duplicate stub-client and decision factories
  into shared helper file (arcjet-ai/test/_shared/stub-client.ts). Refactored
  protect-action.test.ts, protect-tool.test.ts, and generate-text.test.ts
  to import from the shared helper instead of duplicating the code. This
  eliminates subtle drift (e.g., missing DecisionDeny type annotation in
  protect-action.test.ts). Test glob pattern (test/*.test.ts) does not
  pick up _shared/ directory, so helper file is not executed as a test.

All 47 tests pass. Lint, typecheck, and format checks all pass.
